backend/pyproject.toml
[project]
name = "paper-review-backend"
version = "0.1.0"
description = "FastAPI backend for Paper Review AI"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy>=2.0.30",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.9",
    "openai>=1.50.0",
    "python-dotenv>=1.0.1",
    "email-validator>=2.0.0",
]

[tool.uv]
managed = true
